+++ packages/fxa-auth-server/lib/routes/oauth.js
@@ -21,7 +21,6 @@ const error = require('../error');
 const oauthRouteUtils = require('./utils/oauth');
 
 module.exports = (log, config, oauthdb, db, mailer, devices) => {
-
   const OAUTH_DISABLE_NEW_CONNECTIONS_FOR_CLIENTS = new Set(
     config.oauth.disableNewConnectionsForClients || []
   );